input changed to field. Focus style
