+++ src/routes/+page.svelte
@@ -1,8 +1,7 @@
 <script>
-	import AtlButton from "$lib/components/atl-button.svelte";
-	import AtlCheckbox from "$lib/components/atl-checkbox.svelte";
-	import AtlInput from "$lib/components/atl-input.svelte";
-
+	import AtlButton from '$lib/components/atl-button.svelte';
+	import AtlCheckbox from '$lib/components/atl-checkbox.svelte';
+	import AtlField from '$lib/components/atl-field.svelte';
 </script>
 
 <h1>Welcome to your library project</h1>
@@ -11,8 +10,7 @@
 
 <AtlButton>buton</AtlButton>
 <AtlCheckbox>checkbox</AtlCheckbox>
-<AtlInput/>
+<AtlField />
 
 <style>
-	
 </style>
